refactor(backend): extract artists controller and update routes
